refactor: adapt to refresh rates in mHz
Disman's API was changed to communicate the refresh rates of an output in mHz
integral values instead of floating Hz values. Adapt to this change.

+++ kcm/output_model.h
@@ -139,7 +139,7 @@ private:
     int refreshRateIndex(const Disman::OutputPtr& output) const;
     QVariantList resolutionsStrings(const Disman::OutputPtr& output) const;
     QVector<QSize> resolutions(const Disman::OutputPtr& output) const;
-    QVector<float> refreshRates(const Disman::OutputPtr& output) const;
+    QVector<int> refreshRates(const Disman::OutputPtr& output) const;
 
     bool positionable(const Output& output) const;
